BackEnd/loginmicroservices/src/main/java/com/rmit/sept/bk_loginservices/web/BookController.java
package com.rmit.sept.bk_loginservices.web;

import com.rmit.sept.bk_loginservices.model.Book;
import com.rmit.sept.bk_loginservices.repositories.BookRepository;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@CrossOrigin(origins = "http://localhost:3000")
@RequestMapping("/api/v1/books")
public class BookController {

  @Autowired
  private BookRepository bookRepository;

  // get all books
  @GetMapping("/all")
  public Iterable<Book> getAllBooks() {
    return bookRepository.findAll();
  }

}
